% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{as_gp3ml_data}
\alias{as_gp3ml_data}
\title{Extract model-ready data from a gp3ml handoff object}
\usage{
as_gp3ml_data(x, ...)
}
\arguments{
\item{x}{A \code{gp3ml_handoff} or \code{gp3ml_handoff_bundle}.}

\item{...}{Reserved for future adapters.}
}
\value{
A data frame.
}
\description{
Extract model-ready data from a gp3ml handoff object
}
