% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-aware-resampling.R
\name{audit_gazepoint_group_folds}
\alias{audit_gazepoint_group_folds}
\title{Aggregate leakage audits across group-aware folds}
\usage{
audit_gazepoint_group_folds(x)
}
\arguments{
\item{x}{A \code{gazepoint_group_folds} object.}
}
\value{
An object of class \code{gazepoint_group_folds_audit}.
}
\description{
Aggregate leakage audits across group-aware folds
}
