% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance-profile.R
\name{audit_gp3ml_governance_profile}
\alias{audit_gp3ml_governance_profile}
\title{Audit a governance-evidence profile}
\usage{
audit_gp3ml_governance_profile(profile)
}
\arguments{
\item{profile}{Governance profile.}
}
\value{
A \code{gp3ml_governance_profile_audit}.
}
\description{
Audit a governance-evidence profile
}
