% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance-profile.R
\name{create_gp3ml_governance_profile}
\alias{create_gp3ml_governance_profile}
\title{Create a governance-evidence profile}
\usage{
create_gp3ml_governance_profile(
  evidence,
  framework = c("gp3ml-native", "NIST-AI-RMF-1.0", "ISO-23894-oriented",
    "ISO-42001-oriented")
)
}
\arguments{
\item{evidence}{Named list of gp3ml evidence objects.}

\item{framework}{Governance crosswalk.}
}
\value{
A \code{gp3ml_governance_profile}.
}
\description{
Create a governance-evidence profile
}
