% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance-metrics.R
\name{gazepoint_regression_metrics}
\alias{gazepoint_regression_metrics}
\title{Regression metrics}
\usage{
gazepoint_regression_metrics(truth, prediction)
}
\arguments{
\item{truth}{Observed numeric outcome values.}

\item{prediction}{Predicted numeric outcome values.}
}
\description{
Regression metrics
}
