% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-engines.R
\name{gp3ml_available_engines}
\alias{gp3ml_available_engines}
\title{List available model engines}
\usage{
gp3ml_available_engines()
}
\value{
A data frame listing supported model-engine names and whether each optional engine is currently available.
}
\description{
List available model engines
}
\examples{
gp3ml_available_engines()
}
