% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-governance.R
\name{gp3ml_prohibited_uses}
\alias{gp3ml_prohibited_uses}
\title{Prohibited gp3ml uses}
\usage{
gp3ml_prohibited_uses()
}
\value{
A character vector of prohibited use descriptions.
}
\description{
Prohibited gp3ml uses
}
