% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal-prediction.R
\name{plot.gp3ml_conformal_coverage}
\alias{plot.gp3ml_conformal_coverage}
\title{Plot conformal coverage}
\usage{
\method{plot}{gp3ml_conformal_coverage}(x, ...)
}
\arguments{
\item{x}{A \code{gp3ml_conformal_coverage}.}

\item{...}{Additional arguments passed to \code{graphics::barplot()}.}
}
\description{
Plot conformal coverage
}
