% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-shift.R
\name{plot.gp3ml_dataset_shift_audit}
\alias{plot.gp3ml_dataset_shift_audit}
\title{Plot a dataset shift audit}
\usage{
\method{plot}{gp3ml_dataset_shift_audit}(x, ...)
}
\arguments{
\item{x}{A \code{gp3ml_dataset_shift_audit}.}

\item{...}{Additional arguments passed to \code{graphics::barplot()}.}
}
\description{
Plot a dataset shift audit
}
