% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-capabilities.R
\name{plot.gp3ml_engine_capabilities}
\alias{plot.gp3ml_engine_capabilities}
\title{Plot gp3ml engine availability}
\usage{
\method{plot}{gp3ml_engine_capabilities}(x, ...)
}
\arguments{
\item{x}{Engine-capability table.}

\item{...}{Additional graphical parameters.}
}
\description{
Plot gp3ml engine availability
}
