% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-provenance.R
\name{plot.gp3ml_environment_comparison}
\alias{plot.gp3ml_environment_comparison}
\title{Plot an environment comparison}
\usage{
\method{plot}{gp3ml_environment_comparison}(x, ...)
}
\arguments{
\item{x}{An environment comparison.}

\item{...}{Additional arguments passed to \code{graphics::barplot()}.}
}
\description{
Plot an environment comparison
}
