% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance-profile.R
\name{plot.gp3ml_governance_profile_audit}
\alias{plot.gp3ml_governance_profile_audit}
\title{Plot a governance profile audit}
\usage{
\method{plot}{gp3ml_governance_profile_audit}(x, ...)
}
\arguments{
\item{x}{Governance profile audit.}

\item{...}{Additional arguments passed to \code{graphics::barplot()}.}
}
\description{
Plot a governance profile audit
}
