% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-artifacts.R
\name{plot.gp3ml_model_artifact_validation}
\alias{plot.gp3ml_model_artifact_validation}
\title{Plot model-artifact validation}
\usage{
\method{plot}{gp3ml_model_artifact_validation}(x, ...)
}
\arguments{
\item{x}{A \code{gp3ml_model_artifact_validation}.}

\item{...}{Additional arguments passed to \code{graphics::barplot()}.}
}
\description{
Plot model-artifact validation
}
