% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness-stability.R
\name{plot.gp3ml_model_robustness_audit}
\alias{plot.gp3ml_model_robustness_audit}
\title{Plot a robustness audit}
\usage{
\method{plot}{gp3ml_model_robustness_audit}(x, ...)
}
\arguments{
\item{x}{A \code{gp3ml_model_robustness_audit}.}

\item{...}{Additional arguments passed to \code{graphics::barplot()}.}
}
\description{
Plot a robustness audit
}
