% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-plan.R
\name{plot.gp3ml_plan_deviation_audit}
\alias{plot.gp3ml_plan_deviation_audit}
\title{Plot analysis-plan deviations}
\usage{
\method{plot}{gp3ml_plan_deviation_audit}(x, ...)
}
\arguments{
\item{x}{A \code{gp3ml_plan_deviation_audit}.}

\item{...}{Additional arguments passed to \code{graphics::barplot()}.}
}
\description{
Plot analysis-plan deviations
}
