% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-workflow.R
\name{plot.gp3ml_research_bundle_validation}
\alias{plot.gp3ml_research_bundle_validation}
\title{Plot research-bundle validation}
\usage{
\method{plot}{gp3ml_research_bundle_validation}(x, ...)
}
\arguments{
\item{x}{A \code{gp3ml_research_bundle_validation}.}

\item{...}{Additional graphical parameters.}
}
\description{
Plot research-bundle validation
}
