% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro-crate.R
\name{plot.gp3ml_ro_crate_validation}
\alias{plot.gp3ml_ro_crate_validation}
\title{Plot RO-Crate validation}
\usage{
\method{plot}{gp3ml_ro_crate_validation}(x, ...)
}
\arguments{
\item{x}{A \code{gp3ml_ro_crate_validation}.}

\item{...}{Additional arguments passed to \code{graphics::barplot()}.}
}
\description{
Plot RO-Crate validation
}
