% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision-governance.R
\name{plot.gp3ml_threshold_evaluation}
\alias{plot.gp3ml_threshold_evaluation}
\title{Plot threshold evaluation}
\usage{
\method{plot}{gp3ml_threshold_evaluation}(x, metric = "balanced_accuracy", ...)
}
\arguments{
\item{x}{A \code{gp3ml_threshold_evaluation}.}

\item{metric}{Metric to plot.}

\item{...}{Additional arguments passed to \code{graphics::plot()}.}
}
\description{
Plot threshold evaluation
}
