% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal-prediction.R
\name{predict_gazepoint_interval}
\alias{predict_gazepoint_interval}
\title{Predict conformal regression intervals}
\usage{
predict_gazepoint_interval(object, prediction)
}
\arguments{
\item{object}{A regression \code{gp3ml_conformal_fit}.}

\item{prediction}{Point predictions.}
}
\value{
Data frame with point prediction, lower, and upper limits.
}
\description{
Predict conformal regression intervals
}
