% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling-diagnostics.R
\name{print.gazepoint_fold_diagnostics_validation}
\alias{print.gazepoint_fold_diagnostics_validation}
\title{Print Gazepoint fold-diagnostics validation}
\usage{
\method{print}{gazepoint_fold_diagnostics_validation}(x, ...)
}
\arguments{
\item{x}{A \code{gazepoint_fold_diagnostics_validation} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print Gazepoint fold-diagnostics validation
}
