% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-artifacts.R
\name{restore_gazepoint_model_artifact}
\alias{restore_gazepoint_model_artifact}
\title{Restore a model artifact}
\usage{
restore_gazepoint_model_artifact(artifact)
}
\arguments{
\item{artifact}{A model artifact.}
}
\value{
A restored artifact with an unbundled model where needed.
}
\description{
Restore a model artifact
}
