% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-engines.R
\name{train_gazepoint_classifier}
\alias{train_gazepoint_classifier}
\title{Generic governed binary-classifier training wrapper}
\usage{
train_gazepoint_classifier(data, task, predictors = NULL, engine = "glm", ...)
}
\arguments{
\item{data}{Analysis data used to train the classifier.}

\item{task}{A governed binary-classification task.}

\item{predictors}{Optional character vector of predictor columns.}

\item{engine}{Classification engine name or custom engine.}

\item{...}{Additional arguments passed to \code{fit_gazepoint_model()}.}
}
\description{
Generic governed binary-classifier training wrapper
}
