% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-plan.R
\name{validate_gazepoint_analysis_plan}
\alias{validate_gazepoint_analysis_plan}
\title{Validate an analysis plan}
\usage{
validate_gazepoint_analysis_plan(plan)
}
\arguments{
\item{plan}{A \code{gp3ml_analysis_plan}.}
}
\value{
A validation object.
}
\description{
Validate an analysis plan
}
