% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal-prediction.R
\name{validate_gazepoint_conformal}
\alias{validate_gazepoint_conformal}
\title{Validate a conformal fit}
\usage{
validate_gazepoint_conformal(object)
}
\arguments{
\item{object}{A conformal fit.}
}
\value{
A validation object.
}
\description{
Validate a conformal fit
}
