% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling-diagnostics.R
\name{validate_gazepoint_fold_diagnostics}
\alias{validate_gazepoint_fold_diagnostics}
\title{Validate Gazepoint fold diagnostics}
\usage{
validate_gazepoint_fold_diagnostics(x)
}
\arguments{
\item{x}{A \code{gazepoint_fold_diagnostics} object.}
}
\value{
An object of class \code{gazepoint_fold_diagnostics_validation}.
}
\description{
Validate Gazepoint fold diagnostics
}
