% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-aware-resampling.R
\name{validate_gazepoint_group_folds}
\alias{validate_gazepoint_group_folds}
\title{Validate group-aware Gazepoint resampling folds}
\usage{
validate_gazepoint_group_folds(x)
}
\arguments{
\item{x}{A \code{gazepoint_group_folds} object.}
}
\value{
An object of class \code{gazepoint_group_folds_validation}.
}
\description{
Validate group-aware Gazepoint resampling folds
}
