% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{validate_gazepoint_handoff}
\alias{validate_gazepoint_handoff}
\title{Validate a Gazepoint handoff}
\usage{
validate_gazepoint_handoff(x)
}
\arguments{
\item{x}{A \code{gp3ml_handoff}.}
}
\value{
A \code{gp3ml_handoff_validation} object.
}
\description{
Validate a Gazepoint handoff
}
