% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-aware-splitting.R
\name{validate_gazepoint_ml_split}
\alias{validate_gazepoint_ml_split}
\title{Validate a group-aware Gazepoint holdout split}
\usage{
validate_gazepoint_ml_split(x)
}
\arguments{
\item{x}{An object returned by \code{\link[=split_gazepoint_ml_data]{split_gazepoint_ml_data()}}.}
}
\value{
An object of class \code{gazepoint_ml_split_validation}.
}
\description{
Validate a group-aware Gazepoint holdout split
}
