% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tuning.R
\name{validate_gazepoint_model_tuning}
\alias{validate_gazepoint_model_tuning}
\alias{print.gp3ml_model_tuning_validation}
\title{Validate governed tuning results}
\usage{
validate_gazepoint_model_tuning(x)

\method{print}{gp3ml_model_tuning_validation}(x, ...)
}
\arguments{
\item{x}{A \code{gp3ml_model_tuning} object.}

\item{...}{Additional arguments passed to the print method.}
}
\value{
A \code{gp3ml_model_tuning_validation}.
}
\description{
Validate governed tuning results
}
