% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested-resampling.R
\name{validate_gazepoint_nested_evaluation}
\alias{validate_gazepoint_nested_evaluation}
\alias{print.gp3ml_nested_evaluation_validation}
\title{Validate a nested evaluation}
\usage{
validate_gazepoint_nested_evaluation(x)

\method{print}{gp3ml_nested_evaluation_validation}(x, ...)
}
\arguments{
\item{x}{A \code{gp3ml_nested_evaluation}.}

\item{...}{Additional arguments passed to the print method.}
}
\value{
A \code{gp3ml_nested_evaluation_validation}.
}
\description{
Validate a nested evaluation
}
