% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested-resampling.R
\name{validate_gazepoint_nested_folds}
\alias{validate_gazepoint_nested_folds}
\alias{print.gp3ml_nested_folds_validation}
\title{Validate nested grouped folds}
\usage{
validate_gazepoint_nested_folds(x)

\method{print}{gp3ml_nested_folds_validation}(x, ...)
}
\arguments{
\item{x}{A \code{gp3ml_nested_folds} object.}

\item{...}{Additional arguments passed to the print method.}
}
\value{
A \code{gp3ml_nested_folds_validation} object.
}
\description{
Validate nested grouped folds
}
