% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro-crate.R
\name{validate_gazepoint_ro_crate}
\alias{validate_gazepoint_ro_crate}
\title{Validate a gp3ml RO-Crate-oriented export}
\usage{
validate_gazepoint_ro_crate(path)
}
\arguments{
\item{path}{Crate directory or \code{gp3ml_ro_crate}.}
}
\value{
A validation object.
}
\description{
Validate a gp3ml RO-Crate-oriented export
}
