% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-engine.R
\name{bake_gazepoint_preprocessor}
\alias{bake_gazepoint_preprocessor}
\title{Apply a fitted preprocessing engine}
\usage{
bake_gazepoint_preprocessor(preprocessor, new_data)
}
\arguments{
\item{preprocessor}{A fitted \code{gp3ml_preprocessor} object.}

\item{new_data}{Data to transform using the fitted parameters.}
}
\description{
Apply a fitted preprocessing engine
}
