% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-engines.R
\name{integrate_black_box_model}
\alias{integrate_black_box_model}
\title{Integrate a controlled black-box model engine}
\usage{
integrate_black_box_model(
  name,
  fit_fun,
  predict_fun,
  supports = c("classification", "regression"),
  probability = TRUE,
  metadata = list(),
  safety_declaration
)
}
\arguments{
\item{name}{Unique name for the custom engine.}

\item{fit_fun}{Function that fits the custom engine.}

\item{predict_fun}{Function that generates predictions.}

\item{supports}{Task types supported by the engine.}

\item{probability}{Whether classification probabilities are supported.}

\item{metadata}{Optional engine metadata.}

\item{safety_declaration}{Named logical safety declarations.}
}
\description{
Integrate a controlled black-box model engine
}
